pyproject.toml
[project]
name = "screenspot-trm"
version = "0.1.0"
description = "Screen UI bounding-box prediction using Tiny Recursive Model (TRM) with CLIP"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["deep-learning", "computer-vision", "ui-grounding", "clip", "trm"]

dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # CLIP
    "open-clip-torch>=2.20.0",
    
    # Data processing
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pillow>=10.0.0",
    
    # Configuration
    "pyyaml>=6.0",
    
    # Progress bar
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
wandb = [
    "wandb>=0.15.0",
]

[project.scripts]
train-screenspot = "train:main"
train-screenspot-cached = "train_cached:main"
precompute-embeddings = "precompute_embeddings:main"
eval-screenspot = "evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "data", "utils"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
